Move code for AsyncTlsConnector into a separate module
