+++ kube/src/client/mod.rs
@@ -14,6 +14,11 @@ use crate::{
 };
 
 #[cfg(feature = "ws")]
+mod ws;
+#[cfg(feature = "ws")]
+use ws::AsyncTlsConnector;
+
+#[cfg(feature = "ws")]
 use async_tungstenite::{
     tokio::{connect_async_with_tls_connector, ConnectStream},
     WebSocketStream,
@@ -25,20 +30,6 @@ use http::{self, Request, StatusCode};
 use k8s_openapi::apimachinery::pkg::apis::meta::v1 as k8s_meta_v1;
 use serde::{de::DeserializeOwned, Deserialize};
 use serde_json::{self, Value};
-
-// TODO Newer version of tokio_native_tls re-exports native_tls.
-#[cfg(all(feature = "ws", feature = "ws-native-tls"))]
-use real_native_tls::{Certificate, Identity, TlsConnector};
-#[cfg(all(feature = "ws", feature = "ws-native-tls"))]
-use tokio_native_tls::TlsConnector as AsyncTlsConnector;
-
-#[cfg(all(feature = "ws", feature = "ws-rustls-tls"))]
-use std::sync::Arc;
-#[cfg(all(feature = "ws", feature = "ws-rustls-tls"))]
-use tokio_rustls::{
-    rustls::{self, Certificate, ClientConfig},
-    webpki, TlsConnector as AsyncTlsConnector,
-};
 
 use std::convert::{TryFrom, TryInto};
 
@@ -450,111 +441,6 @@ impl TryFrom<Config> for reqwest::ClientBuilder {
         Ok(builder)
     }
 }
-
-#[cfg(all(feature = "ws", feature = "ws-native-tls"))]
-impl TryFrom<Config> for AsyncTlsConnector {
-    type Error = Error;
-    fn try_from(config: Config) -> Result<Self> {
-        let mut builder = TlsConnector::builder();
-        if let Some((identity, identity_password)) = config.identity.as_ref() {
-            builder.identity(
-                Identity::from_pkcs12(identity, identity_password)
-                    .map_err(|e| Error::SslError(format!("{}", e)))?,
-            );
-        }
-        if let Some(ders) = config.root_cert {
-            for der in ders {
-                builder.add_root_certificate(
-                    Certificate::from_der(&der.0).map_err(|e| Error::SslError(format!("{}", e)))?,
-                );
-            }
-        }
-        if config.accept_invalid_certs {
-            builder.danger_accept_invalid_certs(config.accept_invalid_certs);
-        }
-        let connector = builder.build().map_err(|e| Error::SslError(format!("{}", e)))?;
-        Ok(AsyncTlsConnector::from(connector))
-    }
-}
-
-#[cfg(all(feature = "ws", feature = "ws-rustls-tls"))]
-impl TryFrom<Config> for AsyncTlsConnector {
-    type Error = Error;
-    fn try_from(config: Config) -> Result<Self> {
-        use rustls::internal::pemfile;
-        use std::io::Cursor;
-
-        let mut client_config = ClientConfig::new();
-        // This is based on how `reqwest` does
-        if let Some((buf, _)) = config.identity.as_ref() {
-            let (key, certs) = {
-                let mut pem = Cursor::new(buf);
-                let certs = pemfile::certs(&mut pem)
-                    .map_err(|_| Error::SslError("No valid certificate was found".into()))?;
-                pem.set_position(0);
-
-                let mut sk = pemfile::pkcs8_private_keys(&mut pem)
-                    .and_then(|pkcs8_keys| {
-                        if pkcs8_keys.is_empty() {
-                            Err(())
-                        } else {
-                            Ok(pkcs8_keys)
-                        }
-                    })
-                    .or_else(|_| {
-                        pem.set_position(0);
-                        pemfile::rsa_private_keys(&mut pem)
-                    })
-                    .map_err(|_| Error::SslError("No valid private key was found".into()))?;
-
-                if let (Some(sk), false) = (sk.pop(), certs.is_empty()) {
-                    (sk, certs)
-                } else {
-                    return Err(Error::SslError("private key or certificate not found".into()));
-                }
-            };
-
-            client_config
-                .set_single_client_cert(certs, key)
-                .map_err(|e| Error::SslError(format!("{}", e)))?;
-        }
-
-        if let Some(ders) = config.root_cert {
-            for der in ders {
-                client_config
-                    .root_store
-                    .add(&Certificate(der.0))
-                    .map_err(|e| Error::SslError(format!("{}", e)))?;
-            }
-        }
-
-        // TODO Need rustls 0.17.0+. Need to update tokio first.
-        // if config.accept_invalid_certs {
-        //     client_config
-        //         .dangerous()
-        //         .set_certificate_verifier(Arc::new(NoCertificateVerification {}));
-        // }
-
-        Ok(AsyncTlsConnector::from(Arc::new(client_config)))
-    }
-}
-
-// TODO Need rustls 0.17.0+. Need to update tokio first.
-// #[cfg(all(feature = "ws", feature = "ws-rustls-tls"))]
-// struct NoCertificateVerification {}
-//
-// #[cfg(all(feature = "ws", feature = "ws-rustls-tls"))]
-// impl rustls::ServerCertVerifier for NoCertificateVerification {
-//     fn verify_server_cert(
-//         &self,
-//         _roots: &rustls::RootCertStore,
-//         _presented_certs: &[rustls::Certificate],
-//         _dns_name: webpki::DNSNameRef<'_>,
-//         _ocsp: &[u8],
-//     ) -> Result<rustls::ServerCertVerified, rustls::TLSError> {
-//         Ok(rustls::ServerCertVerified::assertion())
-//     }
-// }
 
 // TODO: replace with Status in k8s openapi?
